requirements_ai.txt
openai>=1.0.0
web3>=6.0.0
python-dotenv>=1.0.0
pyyaml>=6.0
